% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{check_cc}
\alias{check_cc}
\title{Check ISO country code}
\usage{
check_cc(country)
}
\arguments{
\item{country}{vector of country codes to check}
}
\value{
logical vector flagging which records are valid ISO country codes
}
\description{
Check ISO country code
}

