% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{dialr-valid}
\alias{dialr-valid}
\alias{is_parsed}
\alias{is_valid}
\alias{is_possible}
\title{Phone number validity checks}
\usage{
is_parsed(x)

is_valid(x)

is_possible(x, detailed = FALSE, type = NULL)
}
\arguments{
\item{x}{A \code{\link{phone}} vector}

\item{detailed}{If \code{FALSE}, \code{is_possible} returns a logical vector. If
\code{TRUE}, it returns a character vector with "IS_POSSIBLE" or the reason for
failure.}

\item{type}{If provided, checks if \code{x} is possible for the given phone number
type.}
}
\description{
These functions check the validity of phone numbers.
}
\details{
For each element of \code{x}:
\itemize{
\item \code{is_parsed(x)}: was this successfully parsed?
\item \code{is_valid(x)}: is this a valid phone number?
\item \code{is_possible(x)}: is this a possible phone number? Return type depends on
\code{detailed}.
}
}
\section{libphonenumber reference}{


\code{is_valid()}: \code{PhoneNumberUtil.isValidNumber()}

\code{is_possible()}: \code{PhoneNumberUtil.isPossibleNumber()},
\code{PhoneNumberUtil.isPossibleNumberWithReason()},
\code{PhoneNumberUtil.isPossibleNumberForType()} or
\code{PhoneNumberUtil.isPossibleNumberForTypeWthReason()} depending on the
provided arguments.
}

\examples{
  x <- phone(c(0, 0123, "0412 345 678", "61412987654", "03 9123 4567", "+12015550123"), "AU")

  is_parsed(x)
  is_valid(x)

  is_possible(x)
  is_possible(x, detailed = TRUE)

  is_possible(x, type = "MOBILE")
  is_possible(x, detailed = TRUE, type = "MOBILE")
}
\seealso{
Other phone functions: \code{\link{dialr-phone}},
  \code{\link{dialr-region}}, \code{\link{dialr-type}},
  \code{\link{dialr}}
}
\concept{phone functions}
