% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{get_cc}
\alias{get_cc}
\title{Get ISO country code}
\usage{
get_cc(country)
}
\arguments{
\item{country}{vector of country names to retrieve ISO country codes for}
}
\value{
vector of ISO country codes (NA where not found)
}
\description{
Get ISO country code
}

