% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-phone_old.R
\name{ph_example}
\alias{ph_example}
\title{Get example phone numbers}
\usage{
ph_example(country, type = NULL, home = NULL, clean = FALSE)
}
\arguments{
\item{country}{Character vector of region codes. If length = 1, the same
region code is used for all phone numbers}

\item{type}{A character vector of phone number types. If \code{NULL} (default),
returns an example "FIXED_LINE" number.}

\item{home}{\link[=dialr-region]{ISO country code} for home country. If provided,
numbers will be formatted for dialing from the home country.}

\item{clean}{Should non-numeric characters be removed? If \code{TRUE}, keeps
numbers}
}
\value{
a character vector of formatted example phone numbers.
}
\description{
\strong{ph_example is soft-deprecated and will be removed in a future release.}
}
\examples{
ph_example("AU")
ph_example("AU", type = "MOBILE")
ph_example("AU", home = "US")
ph_example("AU", clean = TRUE)
}
