% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-phone_old.R
\name{ph_format}
\alias{ph_format}
\title{Format a phone number}
\usage{
ph_format(phone, country, format = "NATIONAL", home = NULL, clean = FALSE)
}
\arguments{
\item{phone}{Character vector of phone numbers to check}

\item{country}{Character vector of region codes. If length = 1, the same
region code is used for all phone numbers}

\item{format}{Phone number format to use.}

\item{home}{\link[=dialr-region]{ISO country code} for home country. If provided,
numbers will be formatted for dialing from the home country.}

\item{clean}{Should non-numeric characters be removed? If \code{TRUE}, keeps
numbers}
}
\value{
a character vector of formatted phone numbers.
}
\description{
\strong{ph_format is soft-deprecated and will be removed in a future release.}
}
\examples{
x <- c(0, 0123, "0404 753 123", "61410123817", "+12015550123")
ph_format(x, "AU")
ph_format(x, "AU", home = "AU") 
}
