% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone_old.R
\name{ph_region}
\alias{ph_region}
\title{Retrieve region for phone number}
\usage{
ph_region(phone, country)
}
\arguments{
\item{phone}{character vector of phone numbers.}

\item{country}{character vector of region codes. If length = 1, the same
region code is used for all phone numbers.}
}
\value{
character vector containing the country code for each phone number.
}
\description{
Retrieve region for phone number
}
