% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-phone_old.R
\name{ph_type}
\alias{ph_type}
\title{Retrieve type for phone number}
\usage{
ph_type(phone, country)
}
\arguments{
\item{phone}{Character vector of phone numbers.}

\item{country}{Character vector of region codes. If length = 1, the same
region code is used for all phone numbers.}
}
\value{
character vector containing the phone number type for each phone number.
}
\description{
\strong{ph_type is soft-deprecated and will be removed in a future release.}
}
