% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-phone_old.R
\name{ph_valid}
\alias{ph_valid}
\title{Check if a phone number is valid for a given region}
\usage{
ph_valid(phone, country, strict = FALSE)
}
\arguments{
\item{phone}{Character vector of phone numbers to check}

\item{country}{Character vector of region codes. If length = 1, the same
region code is used for all phone numbers}

\item{strict}{If strict = FALSE, ph_valid checks if the phone number is valid
to dial within the given region, including correctly specified
international numbers. If strict = TRUE, ph_valid checks if the
phone number is a domestic phone number for the given region.}
}
\value{
logical vector flagging valid and invalid phone numbers
}
\description{
\strong{ph_valid is soft-deprecated and will be removed in a future release.}
}
\examples{
x <- c(0, 0123, "0404 753 123", "61410123817", "+12015550123")
ph_valid(x, "AU")
}
