% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{check_cc}
\alias{check_cc}
\title{Check ISO country code}
\usage{
check_cc(country)
}
\arguments{
\item{country}{vector of ISO country codes}
}
\value{
logical vector flagging which elements are valid ISO country codes
}
\description{
Check ISO country code
}
