% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialr.R
\docType{package}
\name{dialr-package}
\alias{dialr}
\alias{dialr-package}
\title{Phone number validation in R}
\description{
dialr is an R interface to Google's libphonenumber library.
}
\author{
\strong{Maintainer}: Danny Smith \email{danny@gorcha.org}

Other contributors:
\itemize{
  \item The Social Research Centre [copyright holder]
}

}
\keyword{internal}
