% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{get_cc}
\alias{get_cc}
\title{Get ISO country code}
\usage{
get_cc(country)
}
\arguments{
\item{country}{A character vector of country names}
}
\value{
A vector of ISO country codes (\code{NA} where not found)
}
\description{
Get ISO country code
}
