% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoder.R
\name{get_geocode}
\alias{get_geocode}
\title{Phone number geographical information}
\usage{
get_geocode(x, locale = getOption("dialr.locale"), home = NULL,
  strict = FALSE)
}
\arguments{
\item{x}{A \link{phone} vector.}

\item{strict}{Should the validity of the phone number be checked? If \code{TRUE},
invalid phone numbers return \code{""}.}
}
\value{
A character vector of text descriptions for each phone number for the
given language code.
}
\description{
Retrieve geographical information related to a phone number.
}
\section{libphonenumber reference}{


\code{get_geocode()}:
\code{PhoneNumberOfflineGeocoder.getDescriptionForValidNumber()} by
default, or \code{PhoneNumberOfflineGeocoder.getDescriptionForNumber()} if
\code{strict = TRUE}.
}

\examples{
x <- phone(c(0, 0123, "0412 345 678", "61412987654", "03 9123 4567", "+12015550123"), "AU")
get_geocode(x)
get_geocode(x, strict = TRUE)

# Specify a home country
get_geocode(x, home = "AU")
get_geocode(x, home = "US")

# Specify a language
get_geocode(x, locale = "de", home = "DE")
}
